ai_game_backend_code/requirements.txt
python-alipay-sdk==3.1.2
asgiref==3.6.0
channels==4.0.0
Django==4.1.5
djangorestframework==3.14.0
PyJWT==2.6.0
djangorestframework-simplejwt==5.2.2
django-redis==5.2.0
channels_redis==4.0.0
requests==2.28.1
uwsgi==2.0.21
daphne==4.0.0
thrift==0.16.0
django-simpleui==2022.12.18                                                                                                                                       
django-mptt==0.14.0
aliyun_python_sdk_core==2.13.36
aliyun-python-sdk-dysmsapi==2.1.2
pika==1.3.1
